requirements.txt
python-telegram-bot==13.15
requests
gradio==4.26.0
